EE transformers - non deprecated api
